ビジネスロジックをServiceクラスからServiceImplへ移動
